fix id type

+++ src/main/java/com/example/swr/s08/Coder.java
@@ -25,7 +25,7 @@ public class Coder {
     public Coder() {
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
